pyproject.toml
[tool.poetry]
name = "stock-bot"
version = "0.1.0"
description = "Discord bot for monitoring the financial markets"
authors = ["holstt"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"
pandas = "^1.5.3"
yfinance = "^0.2.12"
table2ascii = "^1.1.1"
pydantic = "^1.10.7"
pyyaml = "^6.0"
matplotlib = "^3.8.2"
typeguard = "^4.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
typeCheckingMode = "basic"


